fix some bug classes load

+++ dataset/__init__.py
@@ -109,5 +109,5 @@ def collate_fn(examples):
     pixel_values = torch.stack(
         [example["video"].permute(1, 0, 2, 3) for example in examples]
     )
-    labels = torch.tensor([example["label"] for example in examples])
+    labels = torch.tensor([example["label"] for example in examples],dtype=torch.long)
     return {"pixel_values": pixel_values, "labels": labels}
